refactor(dashboard): address Copilot review on cost-badge mode
- Extract the CostBadgeMode union/default/guard/labels/formatter into a
  plain lib/cost-badge-mode.ts so the store imports a non-React module
  instead of the SidebarCostBadge.tsx component (cleaner dependency
  direction). The component re-exports the primitives for compatibility.
- isCostBadgeMode now uses Object.prototype.hasOwnProperty instead of the
  `in` operator, so inherited keys (toString / constructor / __proto__)
  can't pass validation from a corrupt localStorage value.
- Use the unicode NBSP escape for the placeholder to match the StatusBar
  convention (no literal invisible character in source).
- Add a regression test for the prototype-pollution guard.
